Refactor the logics and extract the unifiedServer logic into a seperate module
