+++ lib/parsedJson.js
@@ -1,3 +1,10 @@
+/*
+* Convert returned objects
+* Built by Favour George
+* 29th July 2019
+*
+*/
+
 const parsedJson = (jsonObject) => JSON.parse(JSON.stringify(jsonObject))
 
 module.exports = {
